% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticAlgorithm.R
\name{add_dfs}
\alias{add_dfs}
\title{Add random dfs to a vector}
\usage{
add_dfs(x, how_many)
}
\arguments{
\item{x}{a vector to add the dfs to}

\item{how_many}{how many dfs?}
}
\value{
Returns \code{c(x, dfs)} with \code{how_many} dfs-elements.
}
\description{
\code{add_dfs} adds random degrees of freedom parameters to a vector.
}
\details{
Read the source code for details.
}
\keyword{internal}
