% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{all_pos_ints}
\alias{all_pos_ints}
\title{Check whether all arguments are stricly positive natural numbers}
\usage{
all_pos_ints(x)
}
\arguments{
\item{x}{a vector containing the elements to be tested.}
}
\value{
Returns \code{TRUE} or \code{FALSE} accordingly.
}
\description{
\code{all_pos_ints} tells whether all the elements in a vector
  are strictly positive natural numbers.
}
