% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_and_correct_data}
\alias{check_and_correct_data}
\title{Check that the data is set correctly and correct if not}
\usage{
check_and_correct_data(data, p)
}
\arguments{
\item{data}{a numeric vector or class \code{'ts'} object containing the data. \code{NA} values are not supported.}

\item{p}{a positive integer specifying the autoregressive order of the model.}
}
\value{
Returns the data as a class 'ts' object.
}
\description{
\code{check_and_correct_data} checks that the data is set correctly and
 throws an error if there is something wrong with the data.
}
\keyword{internal}
