% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_data}
\alias{check_data}
\title{Check that given object contains data}
\usage{
check_data(object)
}
\arguments{
\item{object}{an object to be tested}
}
\value{
Doesn't return anything but throws and error if something is wrong.
}
\description{
\code{check_data} checks that a given object contains data.
}
\keyword{internal}
