% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_gsmar}
\alias{check_gsmar}
\title{Check that given object has class attribute 'gsmar'}
\usage{
check_gsmar(object, object_name = "gsmar")
}
\arguments{
\item{object}{an object to be tested}

\item{object_name}{the name of the tested object}
}
\value{
Doesn't return anything but throws and error if something is wrong.
}
\description{
\code{check_gsmar} checks that the given object has class attribute 'gsmar'.
}
\keyword{internal}
