% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_model}
\alias{check_model}
\title{Check that the argument model is correctly specified.}
\usage{
check_model(model)
}
\arguments{
\item{model}{is "GMAR", "StMAR" or "G-StMAR" model considered? In G-StMAR model the first \code{M1} components
are \emph{GMAR-type} and the rest \code{M2} components are \emph{StMAR-type}.}
}
\value{
Doesn't return anything, but throws and error if something is wrong.
}
\description{
\code{check_model} checks that the argument model is correctly specified.
}
