% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{checkAndCorrectData}
\alias{checkAndCorrectData}
\title{Check the data is set correctly and correct if not}
\usage{
checkAndCorrectData(data, p)
}
\arguments{
\item{data}{a numeric vector or column matrix containing the data. \code{NA} values are not supported.}

\item{p}{a positive integer specifying the order of AR coefficients.}
}
\value{
Returns a numeric column matrix containing the data.
}
\description{
\code{checkAndCorrectData} checks that the data is set correctly and corrects it if not.
 Throws an error if it can't convert the data to the correct form.
}
