% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{checkLogicals}
\alias{checkLogicals}
\title{Check that the logical arguments StMAR and GStMAR don't conflict}
\usage{
checkLogicals(StMAR, GStMAR)
}
\arguments{
\item{StMAR}{an (optional) logical argument stating whether StMAR model should be considered instead of GMAR model. Default is \code{FALSE}.}

\item{GStMAR}{an (optional) logical argument stating whether G-StMAR model should be considered instead of GMAR model. In G-StMAR model the first M1 components
are GMAR-type and the rest M2 components are StMAR-type. Default is \code{FALSE}.}
}
\value{
Doesn't return anything, but throws and error if something is wrong.
}
\description{
\code{checkLogicals} checks that the logical arguments StMAR and GStMAR don't conflict.
}
