% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{checkPM}
\alias{checkPM}
\title{Check p and M are correctly set}
\usage{
checkPM(p, M)
}
\arguments{
\item{p}{a positive integer specifying the order of AR coefficients.}

\item{M}{a positive integer specifying the number of mixture components or regimes.}
}
\value{
Doesn't return anything, but throws and error if something is wrong.
}
\description{
\code{checkPM} checks that the arguments p and M are correctly set.
}
