% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morePlots.R
\name{condmeanPlot}
\alias{condmeanPlot}
\title{Condinional mean plot for GMAR, StMAR, and G-StMAR models}
\usage{
condmeanPlot(gsmar, ...)
}
\arguments{
\item{gsmar}{object of class \code{'gsmar'} created with the function \code{fitGSMAR} or \code{GSMAR}.}

\item{...}{Grapchical parameters passed to FILL IN}
}
\value{
\code{condmeanPlot} only plots to a graphical device and does not return anything. Numerical values
 of the conditional means can be extracted from the model with the dollar sign.
}
\description{
\code{condmeanPlot} plots the in-sample conditional mean and regimewise conditional means of the model
  along with the time series contained in the model (e.g. the time series the model was fitted to).
}
\details{
FILL IN
}
\examples{
\donttest{
# Examples to be written.
}
}
\references{
\itemize{
   \item Galbraith, R., Galbraith, J. 1974. On the inverses of some patterned matrices arising
           in the theory of stationary time series. \emph{Journal of Applied Probability} \strong{11}, 63-71.
   \item Kalliovirta L. (2012) Misspecification tests based on quantile residuals.
           \emph{The Econometrics Journal}, \strong{15}, 358-393.
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2015. Gaussian Mixture Autoregressive model for univariate time series.
           \emph{Journal of Time Series Analysis}, \strong{36}, 247-266.
   \item Meitz M., Preve D., Saikkonen P. 2018. A mixture autoregressive model based on Student's t-distribution.
           arXiv:1805.04010 \strong{[econ.EM]}.
   \item Virolainen S. 2020. A mixture autoregressive model based on Gaussian and Student's t-distribution.	arXiv:2003.05221 [econ.EM].
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{diagnosticPlot}}, \code{\link{fitGSMAR}}, \code{\link{GSMAR}}, \code{\link{quantileResidualTests}},
 \code{\link{quantileResidualPlot}}, \code{\link{simulateGSMAR}}
}
