% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{forecastGMAR}
\alias{forecastGMAR}
\title{forecastGMAR is deprecated}
\usage{
forecastGMAR(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function is deprecated! Use \code{predict.gsmar()} instead!
}
