% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{format_valuef}
\alias{format_valuef}
\title{Function factory for formatting values}
\usage{
format_valuef(digits)
}
\arguments{
\item{digits}{number of digits to use}
}
\value{
Returns a function that takes an atomic vector as its argument
  and returns it formatted to character with \code{digits} decimals.
}
\description{
\code{format_valuef} generates functions that format
  values so that they print out with the desired number of digits.
}
