% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{get_IC}
\alias{get_IC}
\title{Calculate AIC, HQIC and BIC}
\usage{
get_IC(loglik, npars, obs)
}
\arguments{
\item{loglik}{log-likelihood value}

\item{npars}{the number of (freely estimated) parameters in the model.}

\item{obs}{the number of observations with initial values excluded for conditional models.}
}
\value{
Returns a data frame containing the information criteria values.
}
\description{
\code{get_IC} calculates AIC, HQIC and BIC
}
\keyword{internal}
