% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncondMoments.R
\name{get_regime_means}
\alias{get_regime_means}
\title{Calculate regime specific means \eqn{\mu_{m}}}
\usage{
get_regime_means(gsmar)
}
\value{
Returns a length \code{M} vector containing the regime mean \eqn{\mu_{m}} in the m:th element.
}
\description{
\code{get_regime_means} calculates the regime means \eqn{\mu_{m} = \phi_{m,0}/(1-\sum\phi_{i,m})}
  for the given GMAR, StMAR, or G-StMAR model
}
\examples{
# GMAR model
params13 <- c(1.4, 0.88, 0.26, 2.46, 0.82, 0.74, 5.0, 0.68, 5.2, 0.72, 0.2)
gmar13 <- GSMAR(p=1, M=3, params=params13, model="GMAR")
get_regime_means(gmar13)

# StMAR model
params12t <- c(1.38, 0.88, 0.27, 3.8, 0.74, 3.15, 0.8, 100, 3.6)
stmar12t <- GSMAR(p=1, M=2, params=params12t, model="StMAR")
get_regime_means(stmar12t)

# G-StMAR model (similar to the StMAR model above)
params12gs <- c(1.38, 0.88, 0.27, 3.8, 0.74, 3.15, 0.8, 3.6)
gstmar12 <- GSMAR(p=1, M=c(1, 1), params=params12gs, model="G-StMAR")
get_regime_means(gstmar12)
}
\references{
\itemize{
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2015. Gaussian Mixture Autoregressive model for univariate time series.
           \emph{Journal of Time Series Analysis}, \strong{36}, 247-266.
   \item Meitz M., Preve D., Saikkonen P. 2021. A mixture autoregressive model based on Student's t-distribution.
           \emph{Communications in Statistics - Theory and Methods}, doi: 10.1080/03610926.2021.1916531
   \item Virolainen S. 2021. A mixture autoregressive model based on Gaussian and Student's t-distributions.
           Studies in Nonlinear Dynamics & Econometrics, doi: 10.1515/snde-2020-0060
 }
}
\seealso{
\code{\link{cond_moments}}, \code{\link{uncond_moments}}, \code{\link{get_regime_vars}},
 \code{\link{get_regime_autocovs}}

Other moment functions: 
\code{\link{cond_moments}()},
\code{\link{get_regime_autocovs}()},
\code{\link{get_regime_vars}()},
\code{\link{uncond_moments}()}
}
\concept{moment functions}
