% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IE}
\alias{IE}
\title{University of Michigan: inflation expectation time series: IE}
\format{A class 'ts' time series object containing 492 observations.}
\source{
\url{https://fred.stlouisfed.org/series/MICH}
}
\usage{
IE
}
\description{
A dataset containing monthly inflation expectations from 1078I to 2018XII.
}
\keyword{datasets}
