% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M10Y1Y}
\alias{M10Y1Y}
\title{Spread between 10-Year and 1-Year treasury rates: M10Y1Y}
\format{
A class 'ts' time series object containing 468 observations.
}
\source{
\url{https://fred.stlouisfed.org/series/GS10} \url{https://fred.stlouisfed.org/series/GS1}
}
\usage{
M10Y1Y
}
\description{
A dataset containing monthly U.S. interest rate spread between the 10-Year Treasury constant
maturity rate and 1-Year Treasury constant maturity rate from 1982 January to 2020 December.
}
\keyword{datasets}
