% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plotGMAR}
\alias{plotGMAR}
\title{plotGMAR is deprecated}
\usage{
plotGMAR(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function is deprecated! Use \code{diagnosticPlot()} instead!
}
