% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.gsmarpred}
\alias{print.gsmarpred}
\title{Print method for class 'gsmarpred' objects}
\usage{
\method{print}{gsmarpred}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class \code{'gsmarpred'} generated by \code{predict.gsmar}.}

\item{...}{currently not in use.}

\item{digits}{the number of digits to be printed}
}
\description{
\code{print.gsmarpred} is a print method for call 'gsmarpred'
 objects created with \code{predict.gsmar}.
}
\examples{
 \donttest{
 # GMAR-model
 params12 <-  c(0.18, 0.93, 0.01, 0.86, 0.68, 0.02, 0.88)
 gmar12 <- GSMAR(logVIX, 1, 2, params12)
 pred <- predict(gmar12, n_ahead=10, plotRes=FALSE)
 pred
 print(pred, digits=3)
 }
}
