% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.gsmarsum}
\alias{print.gsmarsum}
\title{Print method from objects of class 'gsmarsum'}
\usage{
\method{print}{gsmarsum}(x, ..., digits)
}
\arguments{
\item{x}{object of class 'gsmarsum' generated by \code{summary.gsmar}.}

\item{...}{currently not in use.}

\item{digits}{the number of digits to be printed}
}
\description{
\code{print.gsmarsum} is a print method for objects of class 'gsmarsum'
 created with the summary method \code{summary.gsmar}. Approximate standard errors
 are printed in brackets.
}
