% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticAlgorithm.R
\name{regime_distance}
\alias{regime_distance}
\title{Calculate "distance" between two regimes}
\usage{
regime_distance(regime_pars1, regime_pars2)
}
\arguments{
\item{regime_pars1}{a numeric vector representing a regime.}

\item{regime_pars2}{a numeric vector representing another regime, same length as \code{regime_pars1}}
}
\value{
Returns "distance" between \code{regime_pars1} and \code{regime_pars2}. Values are scaled
  before calculating the "distance". Read the source code for details.
}
\description{
\code{regime_distance} scales each regime parameter and then calculates distance
 between scaled \code{regime_pars1} and \code{regime_pars2}.
}
