% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simudata}
\alias{simudata}
\title{Simulated data}
\format{A numeric vector of length 200.}
\source{
Simulated
}
\usage{
simudata
}
\description{
A dataset containing 200 observations simulated from a GMAR p=1, M=2 process.
}
\keyword{datasets}
