% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{simulateGMAR}
\alias{simulateGMAR}
\title{simulateGMAR is deprecated}
\usage{
simulateGMAR(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function is deprecated! Use \code{simulateGSMAR()} instead!
}
