% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBFF}
\alias{TBFF}
\title{Spread between the 3-month Treasury bill rate and the effective federal funds rate: TBFF}
\format{
A class 'ts' time series object containing 781 observations.
}
\source{
\url{https://fred.stlouisfed.org/series/TB3SMFFM}
}
\usage{
TBFF
}
\description{
A dataset containing the monthly U.S. interest rate spread between the 3-month Treasury bill secondary
market rate and the effective federal funds rate from 1954 July to 2019 July (781 observations).
This series was studied in the empirical application of Virolainen (forthcoming) introducing the
G-StMAR model.
}
\references{
\itemize{
   \item Virolainen S. forthcoming. A mixture autoregressive model based on Gaussian and Student's t-distributions.
     Studies in Nonlinear Dynamics & Econometrics.
 }
}
\keyword{datasets}
