% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uGMAR.R
\docType{package}
\name{uGMAR-package}
\alias{uGMAR}
\alias{uGMAR-package}
\title{uGMAR: Estimate Univariate Gaussian and Student's t Mixture Autoregressive Models}
\description{
\code{uGMAR} is a package for estimating univariate Gaussian mixture autoregressive (GMAR),
  Student's t mixture autoregressive (StMAR), and Gaussian and Student's t mixture autoregressive (G-StMAR) models.
  In addition to unconstrained and constrained estimation, \code{uGMAR} provides tools for quantile residual based model
  diagnostics, forecasting, simulation, and more.

  The readme file or the vignette is a good place to start.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/saviviro/uGMAR/issues}
}

}
\author{
\strong{Maintainer}: Savi Virolainen \email{savi.virolainen@helsinki.fi} (\href{https://orcid.org/0000-0002-5075-6821}{ORCID})

}
