% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uGMAR.R
\docType{package}
\name{uGMAR}
\alias{uGMAR}
\title{uGMAR: Estimate Univariate Gaussian and Student's t Mixture Autoregressive Models}
\description{
\code{uGMAR} is a package for estimating univariate Gaussian mixture autoregressive (GMAR),
  Student's t mixture autoregressive (StMAR), and Gaussian and Student's t mixture autoregressive (G-StMAR) models.
  In addition to unconstrained and constrained estimation, \code{uGMAR} provides tools for quantile residual based model
  diagnostics, forecasting, simulation, and more.

  The readme file or the vignette is a good place to start.
}
