% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VIX}
\alias{VIX}
\title{CBOE Volatility Index: VIX}
\format{A numeric vector containing 252 values.}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
VIX
}
\description{
A dataset containing CBOE Volatility Index from 5th of July 2016 to 4th of July 2017.
All missing values are simply passed and deleted from the data.
}
\keyword{datasets}
