% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VIX10}
\alias{VIX10}
\title{CBOE Volatility Index: VIX}
\format{A numeric vector containing 2516 values.}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
VIX10
}
\description{
A dataset containing daily CBOE Volatility Index from 1th of January 2009
 to 31th of December 2018.
}
\keyword{datasets}
