% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VIX29}
\alias{VIX29}
\title{CBOE Volatility Index: VIX29}
\format{A numeric vector containing 7304 values.}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
VIX29
}
\description{
A dataset containing daily CBOE Volatility Index from 1th of January 1990
 to 31th of December 2018.
}
\keyword{datasets}
