% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VIX5}
\alias{VIX5}
\title{CBOE Volatility Index: VIX5}
\format{A numeric vector containing 1258 values.}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
VIX5
}
\description{
A dataset containing daily CBOE Volatility Index from 1th of January 2014
 to 31th of December 2018.
}
\keyword{datasets}
