% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W10Y2Y}
\alias{W10Y2Y}
\title{Spread between 10-Year and 2-Year treasury Euro area Government bond interest rates: T10Y1Y}
\format{
A class 'ts' time series object containing 731 observations.
}
\source{
\url{https://sdw.ecb.europa.eu/quickview.do?SERIES_KEY=165.YC.B.U2.EUR.4F.G_N_A.SV_C_YM.SRS_10Y_2Y}
}
\usage{
W10Y2Y
}
\description{
A dataset containing weekly Euroa area interest rate spread between the 10-Year and 2-Year
government bond interest rate from 2007 Jan 1 to 2020 Dec 31 (all issuers whose rating is triple A).
The dates in the ts object are approximate and assume 52 week years.
}
\keyword{datasets}
