% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{warn_ar_roots}
\alias{warn_ar_roots}
\title{Warn about near-unit-roots in some regimes}
\usage{
warn_ar_roots(gsmar, tol = 0.005)
}
\arguments{
\item{tol}{if some root is smaller that \code{1 + tol}, a warning is thrown}
}
\value{
Doesn't return anything.
}
\description{
\code{warn_ar_roots} warns if the model contains near-unit-roots in some regimes
}
\details{
Warns if some moduli of the autoregressive polynomial's roots are close to one.
}
\keyword{internal}
