% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{warn_dfs}
\alias{warn_dfs}
\title{Warn about large degrees of freedom paramater values}
\usage{
warn_dfs(object, warn_about = c("derivs", "errors"))
}
\arguments{
\item{object}{an object to be tested}

\item{warn_about}{warn about inaccurate derivatives or standard errors?}
}
\value{
Doesn't return anything but throws a warning if any degrees of freedom parameters have value
  larger than 1000.
}
\description{
\code{warn_dfs} warns if the model contains large degrees of freedom paramater values
  possibly indicating unrealible numerical derivatives.
}
