% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_gsmar}
\alias{check_gsmar}
\title{Check that given object has class attribute 'gsmar'}
\usage{
check_gsmar(object)
}
\arguments{
\item{object}{an object to be tested}
}
\value{
Doesn't return anything but throws and error if something is wrong.
}
\description{
\code{check_gsmar} checks that the given object has class attribute 'gsmar'.
}
