% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{checkAndCorrectData}
\alias{checkAndCorrectData}
\title{Check that the data is set correctly and correct if not}
\usage{
checkAndCorrectData(data, p)
}
\arguments{
\item{data}{a numeric vector class \code{'ts'} object containing the data. \code{NA} values are not supported.}

\item{p}{a positive integer specifying the order of AR coefficients.}
}
\value{
Returns the data as a class 'ts' object.
}
\description{
\code{checkAndCorrectData} checks that the data is set correctly and
 throws an error if there is something wrong with the data.
}
