% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logVIX}
\alias{logVIX}
\title{CBOE Volatility Index: logVIX}
\format{A numeric vector containing 252 observations.}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
logVIX
}
\description{
A dataset containing natural logarithm of daily CBOE Volatility Index from 5th of July 2016
 to 4th of July 2017.
}
\keyword{datasets}
