% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.gsmarpred}
\alias{print.gsmarpred}
\title{Print method for class 'gsmarpred' objects}
\usage{
\method{print}{gsmarpred}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class \code{'gsmarpred'} generated by \code{predict.gsmar}.}

\item{...}{currently not in use.}

\item{digits}{the number of digits to be printed}
}
\description{
\code{print.gsmarpred} is a print method for call 'gsmarpred'
 objects created with \code{predict.gsmar}.
}
