% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.gsmarsum}
\alias{print.gsmarsum}
\title{Print method from objects of class 'gsmarsum'}
\usage{
\method{print}{gsmarsum}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class 'gsmarsum' generated by \code{summary.gsmar}.}

\item{...}{currectly not used.}

\item{digits}{number of digits to be printed}
}
\description{
\code{print.gsmarsum} is a print method for object \code{'gsmarsum'} created
 with \code{summary.gsmar}. Approximate tandard errors are printed in brackets.
}
\examples{
\donttest{
# GMAR model
fit12 <- fitGSMAR(VIX, 1, 2, nCalls=4)
gsmarsum12 <- summary(fit12)
gsmarsum12
print(gsmarsum12, digits=4)
}
}
