% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlot.R
\name{profile_logliks}
\alias{profile_logliks}
\title{Plot profile log-likehoods around the estimates}
\usage{
profile_logliks(gsmar, scale = 0.02, nrows, ncols, precission = 200)
}
\arguments{
\item{gsmar}{object of class \code{'gsmar'} created with the function \code{fitGSMAR} or \code{GSMAR}.}

\item{scale}{a numeric scalar specifying the interval plotted for each estimate: the estimate plus-minus \code{abs(scale*estimate)}.}

\item{nrows}{how many rows should be in the plot-matrix? The default is \code{max(ceiling(log2(nparams) - 1), 1)}.}

\item{ncols}{how many columns should be in the plot-matrix? The default is \code{ceiling(nparams/nrows)}.
Note that \code{nrows*ncols} should not be smaller than the number of parameters.}

\item{precission}{at how many points should each profile log-likelihood be evaluated at?}
}
\value{
Only plots to a graphical device and doesn't return anything.
}
\description{
\code{profile_logliks} plots profile log-likelihoods around the estimates.
}
\details{
The red vertical line points the estimate.

  Be aware that the profile log-likelihood function is subject to a numerical error due to limited float-point
  precission when considering extremely large parameter values, say, overly large degrees freedom estimates (see the
  related example below).
}
\examples{
\donttest{
# GMAR model
fit12 <- fitGSMAR(data=logVIX, p=1, M=2, model="GMAR")
profile_logliks(fit12)

# Non-mixture version of StMAR model
fit11t <- fitGSMAR(logVIX, 1, 1, model="StMAR", ncores=1, ncalls=1)
profile_logliks(fit11t)

# Restricted G-StMAR-model
fit12gsr <- fitGSMAR(logVIX, 1, M=c(1, 1), model="G-StMAR",
 restricted=TRUE)
profile_logliks(fit12gsr)

# Extremely large degrees of freedom numerical error demonstration
fit12t <- fitGSMAR(logVIX, 1, 2, model="StMAR", ncores=1,
 ncalls=1, seeds=1)
profile_logliks(fit12t, scale=0.00001)
# See the last figure? Surface of the profile log-likelihood function
# should be flat around that large degrees of freedom!
}
}
\references{
\itemize{
   \item Galbraith, R., Galbraith, J. 1974. On the inverses of some patterned matrices arising
           in the theory of stationary time series. \emph{Journal of Applied Probability} \strong{11}, 63-71.
   \item Kalliovirta L. (2012) Misspecification tests based on quantile residuals.
           \emph{The Econometrics Journal}, \strong{15}, 358-393.
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2015. Gaussian Mixture Autoregressive model for univariate time series.
           \emph{Journal of Time Series Analysis}, \strong{36}, 247-266.
   \item Meitz M., Preve D., Saikkonen P. 2018. A mixture autoregressive model based on Student's t-distribution.
           arXiv:1805.04010 \strong{[econ.EM]}.
   \item There are currently no published references for the G-StMAR model, but it's a straightforward generalization with
           theoretical properties similar to the GMAR and StMAR models.
 }
}
\seealso{
\code{\link{quantileResidualPlot}}, \code{\link{diagnosticPlot}}, \code{\link{fitGSMAR}}, \code{\link{GSMAR}},
 \code{\link{quantileResidualTests}}, \code{\link{simulateGSMAR}}
}
