% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{quantileResidualPlot}
\alias{quantileResidualPlot}
\title{DEPRECATED, USE \code{quantile_residual_plot} INSTEAD! Plot quantile residual time series and histogram}
\usage{
quantileResidualPlot(gsmar)
}
\arguments{
\item{gsmar}{a class 'gsmar' object, typically generated by \code{fitGSMAR} or \code{GSMAR}.}
}
\value{
Only plots to a graphical device and doesn't return anything.
}
\description{
\code{quantileResidualsPlot} plots quantile residual time series and histogram.
 DEPRECATED, USE \code{quantile_residual_plot} INSTEAD!
}
\details{
DEPRECATED, USE \code{quantile_residual_plot} INSTEAD!
}
\references{
\itemize{
   \item Galbraith, R., Galbraith, J. 1974. On the inverses of some patterned matrices arising
           in the theory of stationary time series. \emph{Journal of Applied Probability} \strong{11}, 63-71.
   \item Kalliovirta L. (2012) Misspecification tests based on quantile residuals.
           \emph{The Econometrics Journal}, \strong{15}, 358-393.
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2015. Gaussian Mixture Autoregressive model for univariate time series.
           \emph{Journal of Time Series Analysis}, \strong{36}(2), 247-266.
   \item Meitz M., Preve D., Saikkonen P. 2023. A mixture autoregressive model based on Student's t-distribution.
         \emph{Communications in Statistics - Theory and Methods}, \strong{52}(2), 499-515.
   \item Virolainen S. 2022. A mixture autoregressive model based on Gaussian and Student's t-distributions.
         Studies in Nonlinear Dynamics & Econometrics, \strong{26}(4) 559-580.
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{diagnostic_plot}}, \code{\link{fitGSMAR}}, \code{\link{GSMAR}},
 \code{\link{quantile_residual_tests}}, \code{\link{simulate.gsmar}}
}
