% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{quantileResidualTests}
\alias{quantileResidualTests}
\title{DEPRECATED, USE \code{quantile_residual_tests} INSTEAD! Quantile residual tests for GMAR, StMAR , and G-StMAR models}
\usage{
quantileResidualTests(
  gsmar,
  lags_ac = c(1, 3, 6, 12),
  lags_ch = lags_ac,
  nsimu = 1,
  print_res = TRUE,
  lagsAC = NULL,
  lagsCH = NULL,
  printRes = NULL
)
}
\arguments{
\item{gsmar}{object of class \code{'gsmar'} created with the function \code{fitGSMAR} or \code{GSMAR}.}

\item{lags_ac}{a numeric vector of positive integers specifying the lags for which autocorrelation is tested.}

\item{lags_ch}{a numeric vector of positive integers specifying the lags for which conditional heteroscedasticity
is tested.}

\item{nsimu}{a positive integer specifying to how many simulated observations the covariance matrix Omega
(see Kalliovirta (2012)) should be based on. If smaller than data size, then omega will be based on the
given data and not on simulated data. Having the covariance matrix omega based on a large simulated sample
might improve the tests size properties.}

\item{print_res}{a logical argument defining whether the results should be printed or not.}

\item{lagsAC}{deprecated! Use \code{lags_ac} instead.}

\item{lagsCH}{deprecated! Use \code{lags_ch} instead.}

\item{printRes}{deprecated! Use \code{print_res} instead.}
}
\value{
Returns an object of class \code{'qrtest'} containing the test results in data frames. In the cases
  of autocorrelation and conditional heteroscedasticity tests, the returned object also contains the
  associated individual statistics and their standard errors, discussed in \emph{Kalliovirta (2012)} at
  the pages 369-370.
}
\description{
\code{quantileResidualTests} performs quantile residual tests for GMAR, StMAR,
 and G-StMAR models, testing normality, autocorrelation, and conditional heteroscedasticity
 of the quantile residuals. DEPRECATED, USE \code{quantile_residual_tests} INSTEAD!
}
\details{
DEPRECATED! USE \code{quantile_residual_tests} INSTEAD!

  For a correctly specified GSMAR model employing the maximum likelihood estimator, the quantile residuals
  are asymptotically independent with standard normal distribution. They can hence be used in a similar
  manner to conventional Pearson's residuals. For more details about quantile residual based diagnostics,
  and in particular, about the quantile residual tests, see the cited article by \emph{Kalliovirta (2012)}.
}
\section{Suggested packages}{

  Install the suggested package "gsl" for faster evaluations in the cases of StMAR and G-StMAR models.
  For large StMAR and G-StMAR models with large data, the evaluations may take significantly long time
  without the package "gsl".
}

\references{
\itemize{
   \item Galbraith, R., Galbraith, J. 1974. On the inverses of some patterned matrices arising
           in the theory of stationary time series. \emph{Journal of Applied Probability} \strong{11}, 63-71.
   \item Kalliovirta L. (2012) Misspecification tests based on quantile residuals.
           \emph{The Econometrics Journal}, \strong{15}, 358-393.
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2015. Gaussian Mixture Autoregressive model for univariate time series.
           \emph{Journal of Time Series Analysis}, \strong{36}, 247-266.
   \item Meitz M., Preve D., Saikkonen P. 2018. A mixture autoregressive model based on Student's t-distribution.
           arXiv:1805.04010 \strong{[econ.EM]}.
   \item Virolainen S. 2020. A mixture autoregressive model based on Gaussian and Student's t-distribution.	arXiv:2003.05221 [econ.EM].
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{fitGSMAR}}, \code{\link{GSMAR}}, \code{\link{diagnostic_plot}},
 \code{\link{predict.gsmar}}, \code{\link{get_test_Omega}},
}
