% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uGMAR.R
\docType{package}
\name{uGMAR}
\alias{uGMAR}
\alias{uGMAR-package}
\title{uGMAR: Estimate Univariate Gaussian or Student's t Mixture Autoregressive Model}
\description{
\code{uGMAR} is a package for estimating univariate Gaussian Mixture Autoregressive (GMAR),
  Student's t Mixture Autoregressive (StMAR) and Gaussian and Student's t Mixture Autoregressive (G-StMAR) models.
  It provides tool for quantile residuals tests, graphical diagnostics, forecasting and simulation. Applying
  general linear constraints to the autoregressive parameters, or restricting them to be the same for all regimes
  is supported.

  Many of the functions documented are not exported but for internal use only. The readme file is a good place to start,
  and the vignette might be useful too.
}
